chore(types): finish the mypy ratchet — zero disabled error codes (#37) (#59)
Re-enable the last three deferred codes and fix all 54 errors:

- no-any-return (20): numpy/scipy scalars wrapped in float()/int()
  (float.__pow__ returns Any in typeshed; numpy 2.4 shape elements are
  Any); JSON loads and array returns cast(...) to precise targets;
  eval._band_mask return tightened to NDArray[np.bool_].
- type-arg (24): bare generics parameterized with the real types
  (spec: dict[str, Any] matching the FilterSpec.model_validate
  precedent; report_pdf._add_summary_page gets dict[str, list[Path]]).
- no-untyped-call (10): all genuinely third-party (schemdraw element
  constructors, matplotlib PdfPages.savefig) — targeted two-module
  [[tool.mypy.overrides]] with the libraries named, not per-line
  ignores.

The global disable_error_code list is now EMPTY — since #30 turned
mypy on (194 errors), correctness was enforced and completeness
deferred; both are enforced from here on.

Also: fix test_runs_headless (#58) to delete DISPLAY and render rather
than asserting the environment is headless — it failed on any dev
machine with X running.

Closes #37.

+++ packages/rf-mcp-common/src/rf_mcp_common/units.py
@@ -44,7 +44,7 @@ def db(linear: float) -> float:
 
 def lin(db_value: float) -> float:
     """dB → linear power ratio."""
-    return 10.0 ** (db_value / 10.0)
+    return float(10.0 ** (db_value / 10.0))
 
 
 def w_to_dbm(power_w: float) -> float:
@@ -56,4 +56,4 @@ def w_to_dbm(power_w: float) -> float:
 
 def dbm_to_w(dbm: float) -> float:
     """dBm → power in watts."""
-    return 10.0 ** (dbm / 10.0) / 1000.0
+    return float(10.0 ** (dbm / 10.0)) / 1000.0
